uint16_t for ADEG LUT

+++ software/src/applets/ADEG.h
@@ -29,7 +29,7 @@ public:
 
         // Inline precomputed LUT for monotonic attack/decay time mapping.
         // Index 0 gives instant response; indices 1..255 map to 1..33333 ticks.
-        static constexpr uint32_t kStageTicks[HEM_ADEG_MAX_VALUE + 1] = {
+        static constexpr uint16_t kStageTicks[HEM_ADEG_MAX_VALUE + 1] = {
             0, 1, 2, 5, 8, 13, 18, 25, 33, 41, 51, 62, 74, 86, 100, 115,
             131, 148, 166, 185, 205, 226, 248, 271, 295, 320, 346, 373, 401, 430, 461, 492,
             524, 558, 592, 627, 664, 701, 739, 779, 819, 861, 903, 947, 992, 1037, 1084, 1131,
